remove definition of BUILD_COMMAND, do install under INSTALL_COMMAND
This way the Makefiles generated by CMake will be able to use all your available cores for building RestFrames. (However many cores you give to 'make -j<N>'.) With your current setup only a single core/process is used for the build step, which can be a bottleneck in the build if RestFrames takes long to build.